+++ easy-ui-react/src/TabNav/TabNavItem.tsx
@@ -1,16 +1,9 @@
-import { useResizeObserver } from "@react-aria/utils";
-import React, {
-  ComponentProps,
-  ElementType,
-  MutableRefObject,
-  ReactNode,
-  useCallback,
-  useEffect,
-  useRef,
-} from "react";
+import React, { ComponentProps, ElementType, ReactNode, useRef } from "react";
 import { useHover } from "react-aria";
 import { classNames } from "../utilities/css";
-import { TabNavContextType, useTabNav } from "./context";
+import { useTabNav } from "./context";
+import { useIndicatorSizing } from "./useIndicatorSizing";
+import { useScrollIntoViewIfNeeded } from "./useScrollIntoViewIfNeeded";
 
 import styles from "./TabNavItem.module.scss";
 
@@ -65,106 +58,4 @@ export function TabNavItem<T extends ElementType = "a">(
       </As>
     </li>
   );
-}
-
-/**
- * Sets the width and position for the indicator (underline) based on the size
- * of the currently selected tab.
- */
-function useIndicatorSizing({
-  isCurrentPage,
-  itemRef,
-  setIndicatorPosition,
-  setIndicatorWidth,
-}: TabNavContextType & {
-  isCurrentPage: boolean;
-  itemRef: MutableRefObject<HTMLElement | null>;
-}) {
-  const handleSizing = useCallback(() => {
-    if (isCurrentPage && itemRef.current) {
-      const $item = itemRef.current;
-      const measurements = getSharedMeasurements($item);
-      setIndicatorWidth(measurements.itemWidth);
-      setIndicatorPosition(measurements.itemPosition);
-    }
-  }, [isCurrentPage, itemRef, setIndicatorPosition, setIndicatorWidth]);
-
-  useEffect(() => {
-    handleSizing();
-  }, [handleSizing]);
-
-  useResizeObserver({
-    ref: itemRef,
-    onResize: handleSizing,
-  });
-}
-
-/**
- * Manages scrolling the tab into view if it's scrolled out of the horizontal
- * line of sight.
- */
-function useScrollIntoViewIfNeeded({
-  isCurrentPage,
-  itemRef,
-}: {
-  isCurrentPage: boolean;
-  itemRef: MutableRefObject<HTMLElement | null>;
-}) {
-  const handleScroll = useCallback(() => {
-    if (isCurrentPage && itemRef.current) {
-      const $item = itemRef.current;
-      const { $nav, navRect, itemPosition, itemWidth, navScrollLeft } =
-        getSharedMeasurements($item);
-
-      const navWidth = navRect.width;
-
-      const itemEndEdge = itemPosition + itemWidth + SCROLL_PADDING;
-      const navEndEdge = navScrollLeft + navWidth;
-
-      const itemStartEdge = itemPosition - SCROLL_PADDING;
-      const navStartEdge = navScrollLeft;
-
-      if (itemEndEdge > navEndEdge) {
-        $nav.scrollTo({
-          left: itemPosition - (navWidth - itemWidth) + SCROLL_PADDING,
-        });
-      }
-
-      if (itemStartEdge < navStartEdge) {
-        $nav.scrollTo({ left: itemPosition - SCROLL_PADDING });
-      }
-    }
-  }, [isCurrentPage, itemRef]);
-
-  useEffect(() => {
-    handleScroll();
-  }, [handleScroll]);
-
-  useResizeObserver({
-    ref: itemRef,
-    onResize: handleScroll,
-  });
-}
-
-function getSharedMeasurements($item: HTMLElement) {
-  const $nav = $item.closest("nav");
-  if (!$nav) {
-    throw new Error("Unable to find parent nav element from tab item");
-  }
-
-  const itemRect = $item.getBoundingClientRect();
-  const navRect = $nav.getBoundingClientRect();
-  const navScrollLeft = $nav.scrollLeft;
-
-  const itemWidth = itemRect.width;
-  const itemPosition = navScrollLeft + (itemRect.x - navRect.x);
-
-  return {
-    $nav,
-    itemRect,
-    navRect,
-    navScrollLeft,
-    itemWidth,
-    itemPosition,
-  };
 }
